requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.13.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.6
Werkzeug==0.16.0
